+++ CASE_MainProject/Assets/Project/Program/GameMain/Effect/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2daba536c9eff13468df869b97f331e1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
