+++ CASE_MainProject/Assets/Project/Program/GameMain/Effect/Materials/煙.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3652dcdc04c9c67489e9a8910648ee25
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
