--- CASE_MainProject/Assets/Project/Program/Title/Gear.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5da898b39ca69414bb3a30005c444ec1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
