+++ CASE_MainProject/Assets/Project/Program/Title/Nothing/TitleAnimator.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0836fa89ff5fa704f87458f16591e730
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
