+++ CASE_MainProject/Assets/Project/Program/Title/Nothing/TitleSelect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 360cbfd10a8493c46abb0438a49bf086
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
