+++ CASE_MainProject/Assets/Project/Program/Title/Nothing/Title_Menu.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2211bd8f03ae5db428cf382e0052d480
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
