--- CASE_MainProject/Assets/Project/Program/Title/TitleAnimator.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0836fa89ff5fa704f87458f16591e730
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
