--- CASE_MainProject/Assets/Project/Program/Title/Title_First.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c06347a68bdb33a44b7a889fc0dfec6e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
