--- CASE_MainProject/Assets/Project/Program/Title/Title_Start.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3c57faaeb0263a74ca5b3680e0e55d47
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
